💄 일정 관리 페이지 반응형 보완
